+++ CHANGES.md
@@ -1,3 +1,7 @@
+#### 0.0.3 / 2022-05-09
+* Added position setting
+* Added defaults when no value set
+
 #### 0.0.2 / 2022-05-06
 * Implemented Settings Link on Plugin Page